move ActionItem to abstract.py
